refactor(NO-TICKET): mypy 부채 152건 0화 - cli 제외 전 모듈, baseline 83
- strategy(54)·live/exchange(36)·ml(33)·backtest/core/data/dashboard
  (29) 의 mypy strict 오류 152건을 어노테이션·캐스트만으로 해소.
  cli.py(83)는 분할 유닛에서 타이핑과 함께 처리 예정 — baseline
  을 235 → 83 으로 하향(래칫 새 바닥).
- 주요 패턴: pandas 스칼라 반환의 bool()/float() 캐스트(기존
  하우스 패턴), bare generic 파라미터화(사용처 기준 정확 타입),
  TYPE_CHECKING 로 optional 의존(lgb·sklearn·optuna·rich) 정밀
  타이핑(lazy-import 관례 유지).
- 행동 인접 변경 2건만 존재, 둘 다 검토 완료: live 엔진 워밍업
  가드 isinstance(Exception) → BaseException (gather(
  return_exceptions=True) 는 BaseException 을 담으므로 엄격히
  더 안전한 내로잉), 엔진 협력자 타입 object → 실클래스
  (OrderManager·TradeValidator·UpbitWebSocketClient top-level
  import — 순환 없음, "object has no attribute" 9건의 근본 수정).
- type: ignore 신규 0건. 4개 영역 병렬 작업 후 통합 검증:
  mypy 83(전부 cli.py)·ruff 0·format clean·434 테스트·
  coverage 62.22%.

Refs: NO-TICKET

+++ src/tradingbot/backtest/report.py
@@ -39,7 +39,7 @@ class BacktestReport:
 
     @property
     def _annualization_factor(self) -> float:
-        return np.sqrt(PERIODS_PER_YEAR.get(self.timeframe, 8766))
+        return float(np.sqrt(PERIODS_PER_YEAR.get(self.timeframe, 8766)))
 
     @property
     def total_trades(self) -> int:
@@ -72,12 +72,12 @@ class BacktestReport:
     @property
     def avg_win(self) -> float:
         wins = [t.pnl for t in self.trades if t.is_win]
-        return np.mean(wins) if wins else 0.0
+        return float(np.mean(wins)) if wins else 0.0
 
     @property
     def avg_loss(self) -> float:
         losses = [t.pnl for t in self.trades if not t.is_win]
-        return np.mean(losses) if losses else 0.0
+        return float(np.mean(losses)) if losses else 0.0
 
     @property
     def profit_factor(self) -> float:
@@ -127,7 +127,7 @@ class BacktestReport:
     @property
     def avg_trade_duration_hours(self) -> float:
         durations = [t.duration for t in self.trades if t.duration is not None]
-        return np.mean(durations) if durations else 0.0
+        return float(np.mean(durations)) if durations else 0.0
 
     def summary(self) -> dict[str, str]:
         """Generate a summary dictionary for display."""
